Update to futures 0.3
This raises the futures dependency from the deprecated 0.1 version to
0.3.
This change aims to make the new Rust 2018 keywords "async" and "await" on
the returned futures usable.

This also removes the `futures-cpupool` dependency and replaces it with
the new `ThreadPool` futures executor.

Additionally, `argonautica-rs` has been switched to the Rust 2018
edition.